ticket 1 created basic styling
I have edited and set up a css sheet with basic styling including google fonts and colours, styling for various h tags and containers such as a top banner, navbar, post container, sidebar and footer. I have tried to include a navbar link that only appears when the blog owner is logged in.

Remaining to do:
Edit the text in the sidebar so it is actually centred- this is proving harder than I thought it would be!?!
Edit the dropdown login text so it does not display the default blue link colour- again, harder than it should be for some reason?!
Make the whole template way more responsive for mobile users- ensuring the 'twitter feed' goes below the posts etc.

Hopefully this is enough for you guys to see what I'm trying to do!